# swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Team Task Manager API",
    "description": "Node.js + Express + MongoDB backend API for team task management. Authentication is handled with Google OAuth and protected routes use the session cookie returned after login.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host (recommended)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local"
    },
    {
      "url": "https://cse341-project-du0g.onrender.com",
      "description": "Render"
    }
  ],
...
  "paths": {
    "/auth/google": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Start Google OAuth login",
        "responses": {
          "302": {
            "description": "Redirect to Google OAuth"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/auth/google/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Google OAuth callback",
        "responses": {
          "302": {
            "description": "Redirect when authentication fails"
          },
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Login successful"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout current user",
        "responses": {
          "200": {
            "description": "Logged out",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Logged out successfully"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get current authenticated user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authenticated": {
                      "type": "boolean",
                      "example": true
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "title"
                },
                "message": {
                  "type": "string",
                  "example": "Title is required"
                }
              }
            }
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "699a31a4ab7b6e21bb36a7f1"
          },
          "name": {
            "type": "string",
            "example": "Alice Tanaka"
          },
          "email": {
            "type": "string",
            "example": "alice@example.com"
          },
          "oauthProvider": {
            "type": "string",
            "example": "google"
          },
          "oauthId": {
            "type": "string",
            "example": "117200000000000000000"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}